old-tests/webapps/XMLHttpRequest/tests/submissions/Opera/responsexml-non-well-formed.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: responseXML non well-formed tests</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function request(content) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open("GET", "resources/status.php&type=text/xml&content=" + content, false)
          client.send(null)
          assert_equals(client.responseXML, null)
        })
      }
      request("<x")
      request("<x><\/x")
      request("<x>&amp<\/x>")
    </script>
  </body>
</html>
